+++ packages/walrus-site-builder-react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @cmdoss/walrus-site-builder-react
 
+## 2.6.0
+
+### Minor Changes
+
+- 4eae3ba: add update metadata modal
+
+### Patch Changes
+
+- Updated dependencies [4eae3ba]
+  - @cmdoss/walrus-site-builder@2.6.0
+  - @cmdoss/file-manager@2.6.0
+
 ## 2.5.0
 
 ### Minor Changes